Create a separate HomeMolecule to make the code more readable
